♻️ refactor: reorganize persistdb files
